refactor: added comments, cleaned up code.
Addressed several issues with the codebase. See
issue #62 for some of the solved details.

- banker.py
-- Most notably added a more robust unit testing
manager. The user can create custom unit tests or
choose from a list of pre-made tests, which are
now detailed as they print to the screen.
-- Moved battleship and tictactoe handling from
handle_data() to their own respective functions.
-- Added validate_name()
-- Added comments, cleaned up code.
- modules.py
-- Moved calculator module from player.py to
modules.py. Streamlined a bit.
- monopoly.py
-- Fixed printing bug where the player's name
was not being printed as player.order instead of
player.name.
- networking.py
-- Added comments, cleaned up code.
- player_class.py
-- Renamed class symbol  from Player to
MonopolyPlayer to clear up some ambiguity. This is
 reflected in cards.py, board.py, monopoly.py
- player.py
-- Added inputting player name during
initialization.
-- Added bal function to get balance from banker.
-- Moved calculator module to modules.py
-- Added comments, cleaned up code.
- screenspace.py
-- Removed print_board()
-- Added whitelist to get_valid_int()
-- Added comments, cleaned up code.
- Deleted testclient.py, testserver.py
- Updated .gitignore