refactor: created Terminal class
screenspace.py
- Creating a Terminal class solves the issue of losing terminal data
on resetting, and allows for easier management of terminal output.
Also, passing functions to Terminals now will be better for
modules which require a lot of output.

style.py
- Removed all get_graphics() calls from all modules, and replaced them
with a reference of the graphics object from the style class.
This improves performance, as get_graphics() was a function call which
was called every time a graphic was needed.