Completed deed viewing module
Fixes #97

Cleaned up output and added pretty banner to deed viewing
Adjusted some variables to allow for more deed information. This should be
the standard when adding new modifiers and effects later down the line,
because most of the code is already written and it is just a matter of
adding the new modifiers and effects where they belong.

Changed monopoly unit testing a little bit.